+++ components/header/index.ts
@@ -1,0 +1,2 @@
+export * from './DefaultHeader';
+export * from './IsBackHeader';
